+++ src/scriptlets/scriptlets-list.ts
@@ -75,6 +75,7 @@ export { trustedDispatchEvent } from './trusted-dispatch-event';
 export { trustedReplaceOutboundText } from './trusted-replace-outbound-text';
 export { preventCanvas } from './prevent-canvas';
 export { trustedReplaceArgument } from './trusted-replace-argument';
+export { preventInnerHTML } from './prevent-innerHTML';
 // redirects as scriptlets
 // https://github.com/AdguardTeam/Scriptlets/issues/300
 export { AmazonApstag } from './amazon-apstag';