+++ src/scriptlets/scriptlets-names-list.ts
@@ -75,6 +75,7 @@ export { trustedDispatchEventNames } from './trusted-dispatch-event';
 export { trustedReplaceOutboundTextNames } from './trusted-replace-outbound-text';
 export { preventCanvasNames } from './prevent-canvas';
 export { trustedReplaceArgumentNames } from './trusted-replace-argument';
+export { preventInnerHTMLNames } from './prevent-innerHTML';
 // redirects as scriptlets
 // https://github.com/AdguardTeam/Scriptlets/issues/300
 export { AmazonApstagNames } from './amazon-apstag';